// src/shared/uiKit/3D/Chips/model/types/types.ts
export enum ChipsNominals {
  CHIP_25_CENTS = 'chip_25_cents',
  CHIP_50_CENTS = 'chip_50_cents',
  CHIP_1_DOLLAR = 'chip_1_dollar',
  CHIP_5_DOLLARS = 'chip_5_dollars',
  CHIP_10_DOLLARS = 'chip_10_dollars',
  CHIP_25_DOLLARS = 'chip_25_dollars',
  CHIP_50_DOLLARS = 'chip_50_dollars',
  CHIP_100_DOLLARS = 'chip_100_dollars',
  CHIP_500_DOLLARS = 'chip_500_dollars',
  CHIP_1000_DOLLARS = 'chip_1000_dollars',
  CHIP_5000_DOLLARS = 'chip_5000_dollars',
  CHIP_10000_DOLLARS = 'chip_10000_dollars',
}